Revert "fix warnings"
This reverts commit 315d33961bbb5349fc3687e8d685252b54b93e3a.

+++ model/correct_0.f90
@@ -21,7 +21,7 @@ MODULE CORRECT_0_MODULE
 
       IMPLICIT NONE
 
-      integer(c_int), intent(in ) :: slo(3),shi(3),lo(3),hi(3)
+      integer     , intent(in   ) :: slo(3),shi(3),lo(3),hi(3)
 
       real(c_real), intent(inout) :: p_g&
          (slo(1):shi(1),slo(2):shi(2),slo(3):shi(3))
